Refactoring: Introduce `wetterdienst.model`, streamline others
