+++ tests/core/test_summary.py
