+++ wetterdienst/core/summarize.py
@@ -10,13 +10,13 @@ from typing import TYPE_CHECKING
 import polars as pl
 from tqdm import tqdm
 
-from wetterdienst.core.timeseries.tools import _ParameterData, extract_station_values
+from wetterdienst.core.util import _ParameterData, extract_station_values
 from wetterdienst.metadata.resolution import Frequency
 from wetterdienst.util.logging import TqdmToLogger
 
 if TYPE_CHECKING:
-    from wetterdienst.core.timeseries.request import TimeseriesRequest
-    from wetterdienst.core.timeseries.result import StationsResult
+    from wetterdienst.model.request import TimeseriesRequest
+    from wetterdienst.model.result import StationsResult
 
 log = logging.getLogger(__name__)
