--- wetterdienst/core/timeseries/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (C) 2018-2025, earthobservations developers.
-# Distributed under the MIT License. See LICENSE for more info.
