+++ wetterdienst/provider/geosphere/observation/api.py
@@ -14,15 +14,15 @@ from zoneinfo import ZoneInfo
 
 import polars as pl
 
-from wetterdienst.core.timeseries.request import TimeseriesRequest
-from wetterdienst.core.timeseries.values import TimeseriesValues
 from wetterdienst.metadata.cache import CacheExpiry
 from wetterdienst.metadata.resolution import Resolution
+from wetterdienst.model.request import TimeseriesRequest
+from wetterdienst.model.values import TimeseriesValues
 from wetterdienst.provider.geosphere.observation.metadata import GeosphereObservationMetadata
 from wetterdienst.util.network import download_file
 
 if TYPE_CHECKING:
-    from wetterdienst.core.timeseries.metadata import ParameterModel
+    from wetterdienst.model.metadata import ParameterModel
 
 log = logging.getLogger(__name__)
