+++ wetterdienst/provider/noaa/ghcn/metadata.py
@@ -2,7 +2,7 @@
 # Distributed under the MIT License. See LICENSE for more info.
 """Metadata for NOAA Global Historical Climatology Network (GHCN) dataset."""
 
-from wetterdienst.core.timeseries.metadata import DATASET_NAME_DEFAULT, build_metadata_model
+from wetterdienst.model.metadata import DATASET_NAME_DEFAULT, build_metadata_model
 
 # translate the above enums to dictionary based model
 NoaaGhcnMetadata = {
